Move TestCsvFileSystemFormatFactory to table runtime
