+++ flink-table/flink-table-runtime-blink/src/test/resources/META-INF/services/org.apache.flink.table.factories.Factory
@@ -13,5 +13,4 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-org.apache.flink.table.planner.factories.TestValuesTableFactory
-org.apache.flink.table.planner.factories.TestFileSourceFactory
+org.apache.flink.table.filesystem.TestCsvFileSystemFormatFactory
